+++ 2_INFORMATIVE_REFERENCE/src/odgs/__init__.py
@@ -1,6 +1,14 @@
 # exposed to user
 from .executive.interceptor import OdgsInterceptor, ProcessBlockedException, SecurityException
+from .executive.exceptions import SoftStopException, DependencyFailedException, ConformanceException
 
-__all__ = ["OdgsInterceptor", "ProcessBlockedException", "SecurityException"]
+__all__ = [
+    "OdgsInterceptor",
+    "ProcessBlockedException",
+    "SecurityException",
+    "SoftStopException",
+    "DependencyFailedException",
+    "ConformanceException",
+]
 
-__version__ = "4.0.0"
+__version__ = "6.0.0"